# docs/architecture/ml-api-spec.yaml
openapi: 3.0.3
info:
  title: VayuGuard ML Service API
  description: >
    Machine Learning inference service for VayuGuard air quality forecasting
    and health risk assessment. Provides AQI forecasts, personalized health
    risk scores, and model management endpoints.
  version: 1.0.0
  contact:
    name: VayuGuard AI/ML Team
    email: ml-team@vayuguard.com
    url: https://vayuguard.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api-staging.vayuguard.com
    description: Staging environment
  - url: https://api.vayuguard.com
    description: Production environment
...
paths:
  /api/forecast:
    post:
      tags: [Forecast]
      summary: Generate AQI forecast
      description: >
        Generate an AQI forecast for a given station and time horizon.
        Supports multiple model types (LSTM, XGBoost, Prophet, Ensemble)
        and configurable forecast horizons up to 168 hours (7 days).
      operationId: generateForecast
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForecastRequest'
            examples:
              standard:
                summary: Standard 48-hour forecast
                value:
                  station_id: 42
                  horizon_hours: 48
                  model_type: ensemble
              quick:
                summary: Quick 24-hour LSTM forecast
                value:
                  station_id: 42
                  horizon_hours: 24
                  model_type: lstm
      responses:
        '200':
          description: Forecast generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastResponse'
              examples:
                success:
                  summary: Successful forecast response
                  value:
                    station_id: 42
                    model_type: ensemble
                    model_version: "2.3.1"
                    generated_at: "2025-01-15T10:30:00Z"
                    forecasts:
                      - forecast_time: "2025-01-15T11:00:00Z"
                        predicted_aqi: 142
                        predicted_category: "Unhealthy for Sensitive Groups"
                        confidence_low: 128
                        confidence_high: 156
                        confidence_score: 0.87
                        predicted_pm25: 55.2
                        predicted_pm10: 88.4
                        predicted_o3: 34.1
                      - forecast_time: "2025-01-15T12:00:00Z"
                        predicted_aqi: 138
                        predicted_category: "Unhealthy for Sensitive Groups"
                        confidence_low: 122
                        confidence_high: 154
                        confidence_score: 0.85
                        predicted_pm25: 52.8
                        predicted_pm10: 85.1
                        predicted_o3: 32.5
                    metadata:
                      total_predictions: 48
                      horizon_hours: 48
                      data_freshness_minutes: 12
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "ValidationError"
                message: "horizon_hours must be between 1 and 168"
                details:
                  field: "horizon_hours"
                  constraint: "1 <= horizon_hours <= 168"
        '404':
          description: Station not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "NotFoundError"
                message: "Station with id 9999 not found"
        '422':
          description: Insufficient data for forecast
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "InsufficientDataError"
                message: "Station 42 has fewer than 24 hours of recent data"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "InferenceError"
                message: "Model inference failed unexpectedly"

    get:
      tags: [Forecast]
      summary: Get cached forecast
      description: >
        Retrieve the most recently generated forecast for a station.
        Returns cached results if available and not expired.
      operationId: getCachedForecast
      parameters:
        - name: station_id
          in: query
          required: true
          schema:
            type: integer
          description: Station ID
        - name: model_type
          in: query
          required: false
          schema:
            type: string
            enum: [lstm, xgboost, prophet, ensemble]
            default: ensemble
          description: Model type to retrieve
        - name: hours
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 168
            default: 24
          description: Number of forecast hours to return
      responses:
        '200':
          description: Cached forecast retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastResponse'
        '404':
          description: No cached forecast available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ForecastRequest:
      type: object
      required:
        - station_id
        - horizon_hours
      properties:
        station_id:
          type: integer
          description: Monitoring station ID
          example: 42
        horizon_hours:
          type: integer
          minimum: 1
          maximum: 168
          description: Forecast horizon in hours (1-168, i.e., up to 7 days)
          example: 48
        model_type:
          type: string
          enum: [lstm, xgboost, prophet, ensemble]
          default: ensemble
          description: ML model type to use for forecasting
        include_confidence:
          type: boolean
          default: true
          description: Whether to include confidence intervals
        confidence_level:
          type: number
          minimum: 0.8
          maximum: 0.99
          default: 0.95
          description: Confidence level for intervals (0.80-0.99)
        include_pollutants:
          type: boolean
          default: true
          description: Whether to include individual pollutant forecasts
    ForecastResponse:
      type: object
      properties:
        station_id:
          type: integer
        model_type:
          type: string
        model_version:
          type: string
        generated_at:
          type: string
          format: date-time
        forecasts:
          type: array
          items:
            $ref: '#/components/schemas/ForecastPoint'
        metadata:
          $ref: '#/components/schemas/ForecastMetadata'
    ForecastPoint:
      type: object
      properties:
        forecast_time:
          type: string
          format: date-time
        predicted_aqi:
          type: integer
          minimum: 0
          maximum: 500
        predicted_category:
          type: string
          enum:
            - Good
            - Moderate
            - Unhealthy for Sensitive Groups
            - Unhealthy
            - Very Unhealthy
            - Hazardous
        confidence_low:
          type: integer
        confidence_high:
          type: integer
        confidence_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        predicted_pm25:
          type: number
          format: float
        predicted_pm10:
          type: number
          format: float
        predicted_o3:
          type: number
          format: float
    ForecastMetadata:
      type: object
      properties:
        total_predictions:
          type: integer
        horizon_hours:
          type: integer
        data_freshness_minutes:
          type: integer
          description: Minutes since the last AQI reading used for this forecast
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type/code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error context